Fix bug where is_alive for thread was returning None if not started. Change to use YggTask instead of old YggThread in DSL driver. Fix bug in handling of model cleanup where the source file is not yet set.

+++ yggdrasil/drivers/ModelDriver.py
@@ -1220,7 +1220,8 @@ class ModelDriver(Driver):
         r"""Remove compile executable."""
         if self.overwrite:
             self.remove_products()
-        if self.function and os.path.isfile(self.model_src):
+        if ((self.function and isinstance(self.model_src, str)
+             and os.path.isfile(self.model_src))):
             assert(os.path.basename(self.model_src).startswith('ygg_'))
             os.remove(self.model_src)
         self.restore_files()
